# enhance: use string view to improve string parsing performance