--- a/core/common/StringTools.h
+++ b/core/common/StringTools.h
@@ -194,7 +194,8 @@ public:
 
     StringViewSplitterIterator() = default;
 
-    StringViewSplitterIterator(StringView str, StringView delimiter) : mStr(str), mDelimiter(delimiter), mPos(0) {
+    StringViewSplitterIterator(StringView str, StringView delimiter, bool skipConsecutiveDelimiters = false)
+        : mStr(str), mDelimiter(delimiter), mPos(0), mSkipConsecutiveDelimiters(skipConsecutiveDelimiters) {
         findNext();
     }
 
@@ -228,6 +229,15 @@ private:
             return;
         }
 
+        if (mSkipConsecutiveDelimiters && !mDelimiter.empty()) {
+            skipDelimiters();
+            if (mPos >= mStr.size()) {
+                mField = {};
+                mPos = StringView::npos;
+                return;
+            }
+        }
+
         size_t end = 0;
         if (mDelimiter.empty()) {
             end = mPos + 1;
@@ -245,6 +255,25 @@ private:
         } else {
             mField = mStr.substr(mPos, end - mPos);
             mPos = end + mDelimiter.size();
+            if (mSkipConsecutiveDelimiters) {
+                skipDelimiters();
+            }
+        }
+    }
+
+    void skipDelimiters() {
+        if (mDelimiter.size() == 1) {
+            // 优化单字符分隔符的情况
+            char delim = mDelimiter[0];
+            while (mPos < mStr.size() && mStr[mPos] == delim) {
+                mPos++;
+            }
+        } else {
+            // 多字符分隔符的情况
+            while (mPos < mStr.size() && mPos + mDelimiter.size() <= mStr.size()
+                   && mStr.substr(mPos, mDelimiter.size()) == mDelimiter) {
+                mPos += mDelimiter.size();
+            }
         }
     }
 
@@ -252,6 +281,7 @@ private:
     StringView mDelimiter;
     StringView mField;
     size_t mPos = StringView::npos;
+    bool mSkipConsecutiveDelimiters = false;
 };
 
 class StringViewSplitter {
@@ -259,15 +289,17 @@ public:
     using value_type = StringView;
     using iterator = StringViewSplitterIterator;
 
-    StringViewSplitter(StringView str, StringView delimiter) : mStr(str), mDelimiter(delimiter) {}
+    StringViewSplitter(StringView str, StringView delimiter, bool skipConsecutiveDelimiters = false)
+        : mStr(str), mDelimiter(delimiter), mSkipConsecutiveDelimiters(skipConsecutiveDelimiters) {}
 
-    iterator begin() const { return iterator(mStr, mDelimiter); }
+    iterator begin() const { return iterator(mStr, mDelimiter, mSkipConsecutiveDelimiters); }
 
     iterator end() const { return iterator(); }
 
 private:
     StringView mStr;
     StringView mDelimiter;
+    bool mSkipConsecutiveDelimiters = false;
 };
 
 template <class T>