--- a/core/host_monitor/common/FastFieldParser.cpp
+++ b/core/host_monitor/common/FastFieldParser.cpp
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include "common/FastFieldParser.h"
+#include "host_monitor/common/FastFieldParser.h"
 
 #include <algorithm>