主机监控中存在大量的字符串分割和解析操作，boost的split算法性能很差。因此通过StringView实现了一个更高性能的字符串分割解析。单测中提交了对应的benchmark

```
=== CPU统计解析性能测试 ===
✅ 结果正确性验证通过
CPU统计解析 (50000 次迭代, 9 行/次):
  Boost::split:    42507791 μs
  FastFieldParser: 15258752 μs
  加速比:          2.79x
  校验和:          1350000 vs 1350000

=== Socket统计解析性能测试 ===
✅ 结果正确性验证通过 (结果: 97)
Socket统计解析 (50000 次迭代):
  Boost::split:    8970002 μs
  FastFieldParser: 1599973 μs
  加速比:          5.61x
  校验和:          4850000 vs 4850000

=== 系统负载解析性能测试 ===
✅ 结果正确性验证通过
系统负载解析 (50000 次迭代):
  Boost::split:    2859560 μs
  FastFieldParser: 619110 μs
  加速比:          4.62x

=== 网络设备统计解析性能测试 ===
✅ 结果正确性验证通过
网络设备解析 (50000 次迭代, 3 行/次):
  Boost::split:    21048802 μs
  FastFieldParser: 12227737 μs
  加速比:          1.72x
  校验和:          259902534000000 vs 259902534000000

=== 单字段访问性能测试 ===
✅ 结果正确性验证通过
单字段访问 (50000 次迭代, 10 字段/次):
  Boost::split:    44532290 μs
  FastFieldParser: 1799524 μs
  加速比:          24.75x
  校验和:          3000000 vs 3000000

=== 进程凭据解析性能测试 ===
✅ 结果正确性验证通过
进程凭据解析 (50000 次迭代):
  Boost::split:    6338722 μs
  FastFieldParser: 497939 μs
  加速比:          12.73x
  校验和:          100200000 vs 100200000

=== 系统Uptime解析性能测试 ===
✅ 结果正确性验证通过 (uptime: 183857.30)
Uptime解析 (50000 次迭代):
  Boost::split:    1559891 μs
  FastFieldParser: 96963 μs
  加速比:          16.09x

=== 批量字段解析性能测试 ===
✅ 结果正确性验证通过 (磁盘状态14字段解析)
批量字段解析 (50000 次迭代, 14 字段):
  逐个调用:        4090363 μs (14次遍历)
  批量调用:        2088197 μs (1次遍历)
  加速比:          1.96x
  校验和:          139013250000 vs 139013250000
  遍历减少:        14.0x (从14次到1次)

=== 进程内存解析性能测试 ===
✅ 结果正确性验证通过
进程内存解析 (50000 次迭代):
  Boost::split:    3131457 μs
  FastFieldParser: 548794 μs
  加速比:          5.71x
  校验和:          4125696000000 vs 4125696000000

=== 进程参数解析性能测试 ===
✅ 结果正确性验证通过 (参数数量: 6)
进程参数解析 (50000 次迭代):
  Boost::split:    4145645 μs
  FastFieldParser: 1749364 μs
  加速比:          2.37x
  校验和:          3750000 vs 3750000
```